Refatoração: Modulariza o código em .h, .c e Makef
